Introduce contao/loupe-bridge
