+++ composer.json
@@ -42,7 +42,7 @@
         "ausi/slug-generator": "^1.1",
         "bacon/bacon-qr-code": "^3.0",
         "cmsig/seal": "^0.12.2",
-        "cmsig/seal-loupe-adapter": "^0.12.2",
+        "cmsig/seal-loupe-adapter": "^0.12.6",
         "cmsig/seal-symfony-bundle": "^0.12.2",
         "contao-components/ace": "^1.8",
         "contao-components/altcha": "^1.0 || ^2.0",
@@ -85,6 +85,7 @@
         "league/flysystem": "^3.7",
         "league/flysystem-bundle": "^3.4",
         "league/flysystem-local": "^3.0",
+        "loupe/loupe": "^0.13.2",
         "matthiasmullie/minify": "^1.3",
         "monolog/monolog": "^3.0",
         "nelmio/cors-bundle": "^2.0.1",
@@ -197,6 +198,7 @@
         "contao/core-bundle": "self.version",
         "contao/faq-bundle": "self.version",
         "contao/listing-bundle": "self.version",
+        "contao/loupe-bridge": "self.version",
         "contao/maker-bundle": "self.version",
         "contao/manager-bundle": "self.version",
         "contao/news-bundle": "self.version",
@@ -224,6 +226,7 @@
             "Contao\\CoreBundle\\": "core-bundle/src/",
             "Contao\\FaqBundle\\": "faq-bundle/src/",
             "Contao\\ListingBundle\\": "listing-bundle/src/",
+            "Contao\\LoupeBridge\\": "loupe-bridge/src/",
             "Contao\\MakerBundle\\": "maker-bundle/src/",
             "Contao\\ManagerBundle\\": "manager-bundle/src/",
             "Contao\\NewsBundle\\": "news-bundle/src/",
@@ -310,6 +313,7 @@
             "contao/core-bundle": "Contao\\CoreBundle\\ContaoManager\\Plugin",
             "contao/faq-bundle": "Contao\\FaqBundle\\ContaoManager\\Plugin",
             "contao/listing-bundle": "Contao\\ListingBundle\\ContaoManager\\Plugin",
+            "contao/loupe-bridge": "Contao\\LoupeBridge\\ContaoManager\\Plugin",
             "contao/maker-bundle": "Contao\\MakerBundle\\ContaoManager\\Plugin",
             "contao/manager-bundle": "Contao\\ManagerBundle\\ContaoManager\\Plugin",
             "contao/news-bundle": "Contao\\NewsBundle\\ContaoManager\\Plugin",
