+++ loupe-bridge/.gitignore
@@ -1,0 +1,3 @@
+# Composer
+/composer.lock
+/vendor/
