+++ loupe-bridge/README.md
@@ -1,0 +1,25 @@
+# Contao Loupe Bridge
+
+[![](https://img.shields.io/packagist/v/contao/loupe-bridge.svg?style=flat-square)](https://packagist.org/packages/contao/manager-bundle)
+[![](https://img.shields.io/packagist/dt/contao/loupe-bridge.svg?style=flat-square)](https://packagist.org/packages/contao/manager-bundle)
+
+The Loupe Bridge provides a pre-configured bridge between Contao core, [SEAL][1] and [Loupe][2].
+
+If you want to use a different search adapter, do not install this package but [your desired SEAL adapter directly instead][5].
+
+Contao is an Open Source PHP Content Management System for people who want a professional website that is easy to
+maintain. Visit the [project website][3] for more information.
+
+## License
+
+Contao is licensed under the terms of the LGPLv3.
+
+## Getting support
+
+Visit the [support page][4] to learn about the available support options.
+
+[1]: https://php-cmsig.github.io/search/
+[2]: https://github.com/loupe-php/loupe
+[3]: https://contao.org
+[4]: https://to.contao.org/support
+[5]: https://packagist.org/providers/cmsig/seal-adapter-implementation
